Split main
