+++ lib/bt_handling.dart
@@ -1,0 +1,157 @@
+import 'package:flutter/foundation.dart'
+    show Uint8List, ValueListenable, kIsWeb;
+import 'package:flutter/material.dart';
+import 'package:universal_ble/universal_ble.dart';
+
+import 'mockble.dart';
+
+// const BT_DEVICE_UUID = "E4:B0:63:81:5B:19";
+const BT_GATT_ID = "a659ee73-460b-45d5-8e63-ab6bf0825942";
+const BT_SERVICE_ID = "e331016b-6618-4f8f-8997-1a2c7c9e5fa3";
+const BT_CHARACTERISTIC_ID = "beb5483e-36e1-4688-b7f5-ea07361b26a8";
+const BT_S2 = "00001800-0000-1000-8000-00805f9b34fb";
+const BT_S3 = "00001801-0000-1000-8000-00805f9b34fb";
+
+class ListNotifier<T> extends ChangeNotifier
+    implements ValueListenable<List<T>> {
+  ListNotifier() : _value = [];
+  final List<T> _value;
+  @override
+  List<T> get value => List.unmodifiable(_value);
+
+  void assign(Iterable<T> it) {
+    _value.clear();
+    _value.addAll(it);
+    notifyListeners();
+  }
+
+  void append(T item) {
+    _value.add(item);
+    notifyListeners();
+  }
+
+  void clear() {
+    _value.clear();
+    notifyListeners();
+  }
+}
+
+class BluetoothHandling {
+  AvailabilityState bluetoothState = AvailabilityState.unknown;
+  final ListNotifier<BleDevice> devices = ListNotifier<BleDevice>();
+  final ValueNotifier<bool> isScanning = ValueNotifier<bool>(false);
+  final ValueNotifier<BleDevice?> selectedDevice =
+      ValueNotifier<BleDevice?>(null);
+  final ListNotifier<BleService> services = ListNotifier<BleService>();
+  late void Function(Uint8List) onNewDataCallback;
+
+  void initializeBluetooth() {
+    UniversalBle.setInstance(MockBlePlatform.instance);
+
+    _updateBluetoothState();
+
+    if (!kIsWeb) {
+      UniversalBle.enableBluetooth(); // this isn't implemented on web
+    }
+
+    UniversalBle.onScanResult = _onScanResult;
+    UniversalBle.onAvailabilityChange = _onBluetoothAvailabilityChanged;
+    UniversalBle.onPairingStateChange = _onPairingStateChange;
+  }
+
+  Future<void> _updateBluetoothState() async {
+    bluetoothState = await UniversalBle.getBluetoothAvailabilityState();
+  }
+
+  void _onScanResult(BleDevice device) {
+    for (var deviceListDevice in devices.value) {
+      if (deviceListDevice.deviceId == device.deviceId) {
+        if (deviceListDevice.name == device.name) {
+          return;
+        }
+      }
+    }
+    devices.append(device);
+  }
+
+  void _onBluetoothAvailabilityChanged(AvailabilityState state) {
+    bluetoothState = state;
+  }
+
+  void stopScan() async {
+    UniversalBle.stopScan();
+    isScanning.value = false;
+  }
+
+  void startScan() async {
+    if (bluetoothState != AvailabilityState.poweredOn) {
+      return;
+    }
+    devices.clear();
+    services.clear();
+    isScanning.value = true;
+    await UniversalBle.startScan(
+      platformConfig: PlatformConfig(
+        web: WebOptions(optionalServices: [
+          BT_SERVICE_ID,
+          BT_CHARACTERISTIC_ID,
+          BT_S2,
+          BT_S3
+        ]),
+      ),
+    );
+  }
+
+  void toggleScan() async {
+    if (isScanning.value) {
+      stopScan();
+    } else {
+      startScan();
+    }
+  }
+
+  void _onPairingStateChange(String deviceId, bool isPaired) {
+    debugPrint('isPaired $deviceId, $isPaired');
+    // _addLog("PairingStateChange - isPaired", isPaired);
+  }
+
+  Future<void> connectToDevice(BleDevice device) async {
+    if (isScanning.value) {
+      stopScan();
+    }
+    try {
+      await UniversalBle.connect(device.deviceId);
+      services.assign(await UniversalBle.discoverServices(device.deviceId));
+      selectedDevice.value = device;
+    } catch (e) {
+      // Error handling can be implemented here
+    }
+  }
+
+  void dispose() {
+    UniversalBle.onScanResult = null;
+    UniversalBle.onAvailabilityChange = null;
+  }
+
+  void subscribeToService(BleService service) async {
+    final deviceId = selectedDevice.value?.deviceId;
+    if (deviceId == null) return;
+
+    // TODO can only subscribe once, otherwise I get "DartError: Exception: Already listening to this characteristic"
+    for (var characteristic in service.characteristics) {
+      if ((characteristic.uuid == BT_CHARACTERISTIC_ID) &&
+          characteristic.properties.contains(CharacteristicProperty.notify)) {
+        UniversalBle.onValueChange =
+            (String deviceId, String characteristicId, Uint8List newData) {
+          // debugPrint('onValueChange $deviceId, $characteristicId, ${hex.encode(value)}');
+          onNewDataCallback(newData);
+        };
+
+        await UniversalBle.setNotifiable(deviceId, service.uuid,
+            characteristic.uuid, BleInputProperty.notification);
+
+        return;
+      }
+    }
+  }
+}
